Allow returning a falsy value or reject Promise
To prevent rendering of a View.  This can surpress jank as well in an
application when changing between views that re-render on async changes.
The problem this fixes is when you render and then wait on async data to
re-render the View.  There will be a momentary flicker unless you
prevent it.